+++ src/lib/shim/src/lib.rs
@@ -33,6 +33,8 @@ mod bindings {
 
 pub mod clone;
 pub mod mmap_box;
+pub mod preempt;
+pub mod reinit_auxvec_random;
 pub mod shimlogger;
 pub mod syscall;
 pub mod tls;
@@ -77,14 +79,51 @@ impl ExecutionContext {
     /// will restore the previous context when dropped.
     pub fn enter(&self) -> ExecutionContextRestorer {
         ExecutionContextRestorer {
-            prev: CURRENT_EXECUTION_CONTEXT.get().replace(*self),
+            prev: self.enter_without_restorer(),
         }
     }
 
     /// Enter this context for the current thread, *without* creating a
     /// restorer. Returns the previous context.
     pub fn enter_without_restorer(&self) -> ExecutionContext {
-        CURRENT_EXECUTION_CONTEXT.get().replace(*self)
+        let current_execution_ctx = CURRENT_EXECUTION_CONTEXT.get();
+        let peeked_prev = current_execution_ctx.get();
+
+        // Potentially enable/disable preemption, being careful that the current
+        // context is set to shadow when calling other internal functions that
+        // require it.
+        let replaced_prev = match (peeked_prev, *self) {
+            (ExecutionContext::Shadow, ExecutionContext::Application) => {
+                // Call preempt::enable before changing context from shadow, so
+                // that it can access shim state.
+                // SAFETY: We only ever switch threads from the shadow execution
+                // context, and we disable preemption when entering the shadow
+                // execution context, so preemption should be disabled for all
+                // other threads in this process.
+                unsafe { preempt::enable() };
+                current_execution_ctx.replace(*self)
+            }
+            (ExecutionContext::Application, ExecutionContext::Shadow) => {
+                // Change context to shadow before calling preempt::disable, so
+                // that it can access shim state.
+                let c = current_execution_ctx.replace(*self);
+                preempt::disable();
+                c
+            }
+            (ExecutionContext::Application, ExecutionContext::Application) => {
+                // No need to actually replace.
+                ExecutionContext::Application
+            }
+            (ExecutionContext::Shadow, ExecutionContext::Shadow) => {
+                // No need to actually replace.
+                ExecutionContext::Shadow
+            }
+        };
+        // It *shouldn't* be possible for the execution context to have changed
+        // out from under us in between the initial peek and the actual
+        // replacement.
+        assert_eq!(peeked_prev, replaced_prev);
+        peeked_prev
     }
 }
 
@@ -476,16 +515,17 @@ fn init_process() {
     log::trace!("Finished shim global init");
 }
 
-/// Wait for "start" event from Shadow, use it to initialize the thread shared
-/// memory block, and optionally to initialize the process shared memory block.
-fn wait_for_start_event(get_initial_working_dir: bool) {
+/// Wait for "start" event from Shadow, using it to set things up for the
+/// current thread, and if `is_first_thread` is true then also for the current
+/// process.
+fn wait_for_start_event(is_first_thread: bool) {
     debug_assert_eq!(ExecutionContext::current(), ExecutionContext::Shadow);
     log::trace!("waiting for start event");
 
     let mut working_dir = [0u8; linux_api::limits::PATH_MAX];
     let working_dir_ptr;
     let working_dir_len;
-    if get_initial_working_dir {
+    if is_first_thread {
         working_dir_ptr = ForeignPtr::from_raw_ptr(working_dir.as_mut_ptr());
         working_dir_len = working_dir.len();
     } else {
@@ -505,7 +545,20 @@ fn wait_for_start_event(get_initial_working_dir: bool) {
         ipc.to_shadow().send(start_req);
         ipc.from_shadow().receive().unwrap()
     });
-    assert!(matches!(res, ShimEventToShim::StartRes));
+    let ShimEventToShim::StartRes(res) = res else {
+        panic!("Unexpected response: {res:?}");
+    };
+    if is_first_thread {
+        // SAFETY: We're ensuring serial execution in this process, and no other
+        // Rust code in this library should have tried accessing the auxiliary
+        // vector yet, so no references should exist.
+        //
+        // WARNING: It's possible that the dynamic linker/loader or constructors
+        // in other dynamically linked libraries *have* run, and that rewriting
+        // this value here will violate safety assumptions in those objects.
+        // Fortunately we haven't observed this in practice.
+        unsafe { reinit_auxvec_random::reinit_auxvec_random(&res.auxvec_random) };
+    }
 
     // SAFETY: shadow should have initialized
     let thread_blk_serialized = unsafe { thread_blk_serialized.assume_init() };
@@ -520,7 +573,7 @@ fn wait_for_start_event(get_initial_working_dir: bool) {
     // TODO: Instead use posix_spawn_file_actions_addchdir_np in the shadow process,
     // which was added in glibc 2.29. Currently this is blocked on debian-10, which
     // uses glibc 2.28.
-    if get_initial_working_dir {
+    if is_first_thread {
         let working_dir = CStr::from_bytes_until_nul(&working_dir).unwrap();
         rustix::process::chdir(working_dir).unwrap();
     }